Fix small issues
