+++ src/main/java/org/waarp/openr66/dao/database/postgres/PostgreSQLTransferDAO.java
@@ -1,10 +1,11 @@
-package org.waarp.openr66.dao.database;
+package org.waarp.openr66.dao.database.postgres;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+import org.waarp.openr66.dao.database.DBTransferDAO;
 import org.waarp.openr66.dao.exception.DAOException;
 
 public class PostgreSQLTransferDAO extends DBTransferDAO {
