+++ src/test/java/org/waarp/openr66/dao/database/mariadb/DBMultipleMonitorMariaDBDAOIT.java
@@ -1,21 +1,21 @@
-package org.waarp.openr66.dao.database.postgreSQL;
+package org.waarp.openr66.dao.database.mariadb;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
 import org.junit.ClassRule;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.waarp.openr66.dao.database.DBBusinessDAOIT;
+import org.testcontainers.containers.MariaDBContainer;
+import org.waarp.openr66.dao.database.DBMultipleMonitorDAOIT;
 
-public class DBBusinessPostgreSQLDAOIT extends DBBusinessDAOIT {
+public class DBMultipleMonitorMariaDBDAOIT extends DBMultipleMonitorDAOIT {
 
-    private String createScript = "postgresql/create.sql";
-    private String populateScript = "postgresql/populate.sql";
-    private String cleanScript = "postgresql/clean.sql";
+    private String createScript = "mysql/create.sql";
+    private String populateScript = "mysql/populate.sql";
+    private String cleanScript = "mysql/clean.sql";
 
     @ClassRule
-    public static PostgreSQLContainer db = new PostgreSQLContainer();
+    public static MariaDBContainer db = new MariaDBContainer();
 
     @Override
     public Connection getConnection() throws SQLException {
