+++ src/test/java/org/waarp/openr66/dao/database/postgres/DBTransferPostgreSQLDAOIT.java
@@ -1,30 +1,29 @@
-package org.waarp.openr66.dao.database.mariaDB;
+package org.waarp.openr66.dao.database.postgres;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
 import org.junit.ClassRule;
-import org.testcontainers.containers.MariaDBContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
 import org.waarp.openr66.dao.database.DBTransferDAO;
 import org.waarp.openr66.dao.database.DBTransferDAOIT;
-import org.waarp.openr66.dao.database.MariaDBTransferDAO;
+import org.waarp.openr66.dao.database.PostgreSQLTransferDAO;
 import org.waarp.openr66.dao.exception.DAOException;
 
-public class DBTransferMariaDBDAOIT extends DBTransferDAOIT {
+public class DBTransferPostgreSQLDAOIT extends DBTransferDAOIT {
 
-    private String createScript = "mysql/create.sql";
-    private String populateScript = "mysql/populate.sql";
-    private String cleanScript = "mysql/clean.sql";
+    private String createScript = "postgresql/create.sql";
+    private String populateScript = "postgresql/populate.sql";
+    private String cleanScript = "postgresql/clean.sql";
 
     @ClassRule
-    public static MariaDBContainer db = new MariaDBContainer();
+    public static PostgreSQLContainer db = new PostgreSQLContainer();
 
     @Override
     public DBTransferDAO getDAO(Connection con) throws DAOException {
-        return new MariaDBTransferDAO(con);
+        return new PostgreSQLTransferDAO(con);
     }
-
 
     @Override
     public Connection getConnection() throws SQLException {
